WIP: Make Completion FSM pluggable
